pyproject.toml
[project]
name = "data-eng-taxi-ibis-dagster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagster>=1.10.12",
    "dagster-duckdb>=0.26.12",
    "dagster-webserver>=1.10.12",
    "ibis-framework[duckdb,pyspark]>=10.5.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "notebook>=7.4.1",
]
